Add support for old style attributes
In order to support most of the cases than can occur with the UI host been the same object with newstyle or old style attributes this changes where needed.

No need to say that having new style attrs for switching on the same host result on a bad behavior. And the same goes into UI host been worl_ctl with new style attrs. Old style attrs can be hosted on the same UI and it works with the worl_ctl as well even if this isn't recommended