strip pipes from skinCluster names
There is a bug if you have non-unique names in your geo. Your node names change to longNames, and the pm.skinCluster command will not accept a name flag that has pipes in it.

You end up with "Invalid deformer name was specified." errors from Maya. Stripping the pipes from the skinCluster name appears to fix this bug.

This fixes the bug on the skinCluster name only. So skinCopy() and importSkin() are fixed. But I do not know if this has any additional bugs related to pipe characters in the filenames of the jSkin or gSkin files. I do know that longNames with pipes work as filenames on MacOS and Linux. I do not know about Windows.